Improved KataGo's performances.
The main program start KataGo's analysis in a smarter way.
Until now, at each move, the following sequence of commands was sent

```
stop
play X XX
kata-analyze interval XXX ownership true
```

The problem is that kata-analyze takes some time to start. Therefore,
a fast sequence of moves would create lag before KataGo could catch up.
This has been solved the following way:

KataGo's object count the number of received inputs and sent output.
Comparing the two gives a **balance value**. If it is negative, it
means that KataGo has undealt with commands. If this is null, then
KataGo's up-to-date. Therefore, we restrict starting analysis
only under this condition.

REMARK: It means that using KataGo.analysis() is DEPRECIATED and
would break this 'up-to-date test' and breaking the analysis.